move monaco editor
