+++ src/app/sections/domain/domain-obj-creation/domain-obj-creation.module.ts
@@ -11,8 +11,8 @@ import { RouterModule } from '@angular/router';
 import { ActionsModule } from '@vality/ng-core';
 
 import { ThriftEditorModule } from '@cc/app/shared/components/thrift-editor';
+import { MonacoEditorModule } from '@cc/components/monaco-editor';
 
-import { MonacoEditorModule } from '../../../monaco-editor';
 import { ThriftViewerModule } from '../../../shared/components/thrift-viewer';
 import { DomainObjCreationComponent } from './domain-obj-creation.component';
