+++ src/app/sections/domain/domain-obj-review/domain-obj-review.module.ts
@@ -10,7 +10,8 @@ import { MatProgressSpinnerModule } from '@angular/material/progress-spinner';
 import { RouterModule } from '@angular/router';
 import { ActionsModule } from '@vality/ng-core';
 
-import { MonacoEditorModule } from '../../../monaco-editor';
+import { MonacoEditorModule } from '@cc/components/monaco-editor';
+
 import { ThriftEditorModule } from '../../../shared/components/thrift-editor';
 import { ThriftViewerModule } from '../../../shared/components/thrift-viewer';
 import { DomainObjReviewComponent } from './domain-obj-review.component';
