+++ src/app/sections/domain/utils.ts
@@ -1,6 +1,6 @@
 import * as uuid from 'uuid/v4';
 
-import { MonacoFile } from '../../monaco-editor';
+import { MonacoFile } from '@cc/components/monaco-editor';
 
 export const toMonacoFile = (content: string): MonacoFile => ({
     uri: `${uuid()}.json`,
