+++ src/components/monaco-editor/bootstrap.ts
