+++ src/components/monaco-editor/monaco-editor.directive.ts
@@ -1,6 +1,6 @@
 import { Directive, ElementRef, Input } from '@angular/core';
 
-import { ComponentChanges } from '../shared';
+import { ComponentChanges } from '../../app/shared';
 import { AbstractMonacoDirective } from './abstarct-monaco.directive';
 import { MonacoFile } from './model';
 import { MonacoEditorService } from './monaco-editor.service';
