+++ src/components/monaco-editor/providers/provider-register.ts
